requirements.txt
Django==4.2.13
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
python-decouple==3.8
django-cors-headers==4.3.1
